literalinclude improvements, nick comments
