+++ source/crud.txt
@@ -18,7 +18,8 @@ Overview
 
 CRUD (Create, Read, Update, Delete) operations allow you to work with
 the data stored inside of MongoDB. CRUD operations fall into two
-categories: read operations and write operations.
+categories: :ref:`read operations <crud-read-operations>` and
+:ref:`write operations <crud-write-operations>`.
 
 .. note::
 
@@ -30,12 +31,14 @@ categories: read operations and write operations.
    operation. A common method of Promise resolution uses
    ``async``/``await`` syntax.
 
+.. _crud-read-operations:
+
 Read Operations
 ---------------
 
 Read operations find and return documents stored within MongoDB.
 There are multiple types of read operation that allow you to
-access data in many different ways. While some read operations focus on
+access data in different ways. While some read operations focus on
 allowing you to view data that exists in the database, other operations
 let you monitor the database for new data that matches certain criteria,
 and still other read operations let you apply a series of
@@ -64,6 +67,7 @@ navigate matched documents.
       :language: javascript
       :start-after: start find crud example
       :end-before: end find crud example
+      :dedent: 4
 
 Aggregate
 ~~~~~~~~~
@@ -84,6 +88,7 @@ collection into a result set for viewing.
       :language: javascript
       :start-after: start aggregate crud example
       :end-before: end aggregate crud example
+      :dedent: 4
 
 Watch / Subscribe
 ~~~~~~~~~~~~~~~~~
@@ -104,6 +109,18 @@ the collection.
       :language: javascript
       :start-after: start watch crud example
       :end-before: end watch crud example
+      :dedent: 4
+
+.. note::
+
+   You can customize read operations to
+   :doc:`skip </crud/read-operations/skip>` a certain number of
+   document at the beginning of the collection of matched documents,
+   :doc:`limit </crud/read-operations/limit>` the number of returned
+   documents, and :doc:`sort <crud/read-operations/sort>` returned
+   documents in a particular order.
+
+.. _crud-write-operations:
 
 Write Operations
 ----------------
@@ -112,7 +129,7 @@ CRUD write operations add, modify, and remove data in MongoDB. Sometimes
 this means putting entirely new data into a collection, but other times
 it involves changing data that already exists, or even removing it
 entirely. There are three distinct write operations: create, update, and
-delete. The driver offers methojds for each variant of write operation:
+delete. The driver offers methods for each variant of write operation:
 
 Insert
 ~~~~~~
@@ -134,6 +151,7 @@ inserted documents that lack such a field.
       :language: javascript
       :start-after: start insert crud example
       :end-before: end insert crud example
+      :dedent: 4
 
 Update / Replace
 ~~~~~~~~~~~~~~~~
@@ -160,6 +178,7 @@ You can configure any update operation to behave as an
       :language: javascript
       :start-after: start update crud example
       :end-before: end update crud example
+      :dedent: 4
 
 Delete
 ~~~~~~
@@ -181,24 +200,18 @@ match the query.
       :language: javascript
       :start-after: start delete crud example
       :end-before: end delete crud example
+      :dedent: 4
+
+.. note::
+
+   You can customize write operations to
+   :doc:`upsert </crud/write-operations/upsert>` instead of update, time
+   out if the method fails to complete within a certain period of time,
+   and write with varying levels of resilience in a replica set.
 
 Compound operations
 -------------------
 
 Some operations combine aspects of read and write operations. Consult
 :doc:`compound operations </crud/compound-operations>` to learn more
-about these hybrid methods.
-
-Customization
--------------
-
-All CRUD operations offer customizable options to tailor behavior for
-your application's specific needs. You can customize read operations to
-:doc:`skip </crud/read-operations/skip>` a certain number of documents
-at the beginning of the collection of matched documents,
-:doc:`limit </crud/read-operations/limit>` the number of returned
-documents, and :doc:`sort <crud/read-operations/sort>` returned
-documents in a particular order. You can customize write operations to
-:doc:`upsert </crud/write-operations/upsert>` instead of update, time
-out if the method fails to complete within a certain period of time,
-and write with varying levels of resilience in a replica set.
+about these hybrid methods.
