+++ source/crud/read-operations/limit.txt
@@ -8,8 +8,9 @@ Use limit to cap the number of documents that can be returned from a
 read operation. Limit functions as a cap on the maximum number of
 documents that the operation can return, but the operation can return
 a smaller number of documents if there are not enough documents present
-to reach the limit. If limit is used with the skip method, the skip will
-apply first and the cap will only apply to the documents left over after
+to reach the limit. If limit is used with the
+:doc:`skip </crud/read-operations/skip>` method, the skip applies
+first and the caponly applies to the documents left over after
 the skip.
 
 Consider a collection containing documents that describe books. To
@@ -17,7 +18,7 @@ insert this data into a collection, run the following operation:
 
 .. code-block:: javascript
 
-   await collection.insertMany([
+   await books.insertMany([
      { "_id": 1, "name": "The Brothers Karamazov", "author": "Dostoyevsky", "length": 824 },
      { "_id": 2, "name": "Les Misérables", "author": "Hugo", "length": 1462 },
      { "_id": 3, "name": "Atlas Shrugged", "author": "Rand", "length": 1088 },
@@ -42,7 +43,8 @@ to return only the 3 longest books:
    const cursor = collection.find(query).sort(sort).limit(limit);
    await cursor.forEach(console.dir);
 
-``find()`` returns the following documents:
+``find()`` returns the following documents sorted from longest length
+to shortest length:
 
 .. code-block:: javascript
