+++ source/crud/read-operations/text.txt
@@ -12,8 +12,8 @@ operator to perform text searches on fields which have a
 Overview
 --------
 
-Text searches let you search a collection of strings for words or
-phrases, instead of matching substrings like
+Text searches let you search a collection of string-valued fields for
+words or phrases, instead of matching substrings like
 `regular expresssions
 <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions>`_,
 also known as RegEx. While regex is very powerful, and MongoDB even
@@ -24,8 +24,11 @@ plurality, case sensitivity, stop words, whitespace, and punctuation in
 multiple languages, all you have to do is query with a search term to
 find documents that contain your term.
 
-You can only create *one* text index per collection, so every text
-search queries all fields used by the index for term matches.
+
+.. admonition::
+
+   You can only create *one* text index per collection, so every text
+   search queries all fields used by the index for term matches.
 
 Examples
 --------
@@ -52,29 +55,11 @@ containing the word "trek". If you want to query using multiple words,
 separate your words with spaces to query for documents that match any of
 the search terms (logical ``OR``).
 
-.. code-block:: javascript
-
-   const database = client.db("sample_mflix");
-   const collection = database.collection("movies");
-
-   const query = { $text: { $search: "trek" } };
-
-   // Return only the `title` of each matched document
-   const projection = {
-      _id: 0,
-      title: 1
-   };
-
-   // find documents based on our query, sort and projection
-   const cursor = collection
-      .find(query)
-      .project(projection);
-
-   // print a message if no documents were found
-   if ((await cursor.count()) == 0) {
-      console.log("No documents found!");
-   }
-   await cursor.forEach(console.dir);
+.. literalinclude:: /code-snippets/crud/startrek.js
+   :language: javascript
+   :start-after: start word text example
+   :end-before: end word text example
+   :dedent: 4
 
 This operation returns the following documents:
 
@@ -97,7 +82,8 @@ This operation returns the following documents:
 Success! The query found every document in the ``movies`` collection
 with a title including the word "trek". Unfortunately, it looks like the
 search included one unintended item: "Trek Nation," which is a movie
-about Star Trek and not a Star Trek movie.
+about Star Trek and not a Star Trek movie. To solve this, we can query
+with a more specific **phrase**.
 
 Query By Phrase
 ~~~~~~~~~~~~~~~
@@ -106,29 +92,11 @@ To make your query more specific, try using the phrase "star trek"
 instead of just the word "trek". To search by phrase, surround your
 multi-word phrase with escaped quotes (``\"<term>\"``):
 
-.. code-block:: javascript
-
-   const database = client.db("sample_mflix");
-   const collection = database.collection("movies");
-
-   const query = { $text: { $search: "\"star trek\"" } };
-
-   // Return only the `title` of each matched document
-   const projection = {
-      _id: 0,
-      title: 1
-   };
-
-   // find documents based on our query, sort and projection
-   const cursor = collection
-      .find(query)
-      .project(projection);
-
-   // print a message if no documents were found
-   if ((await cursor.count()) == 0) {
-      console.log("No documents found!");
-   }
-   await cursor.forEach(console.dir);
+.. literalinclude:: /code-snippets/crud/startrek.js
+   :language: javascript
+   :start-after: start phrase text example
+   :end-before: end phrase text example
+   :dedent: 4
 
 Querying by the phrase "star trek" instead of just the term "trek"
 matches the following documents:
@@ -151,7 +119,8 @@ matches the following documents:
 These results include all movies in the database that contain the phrase
 "star trek", which in this case results in only fictional Star Trek
 movies. Unfortunately, though, this query returned "Star Trek Into
-Darkness", a movie that nobody liked much.
+Darkness", a movie that nobody liked much. To solve this, we can omit
+that document with a **negation**.
 
 Query with Negations
 ~~~~~~~~~~~~~~~~~~~~
@@ -163,29 +132,11 @@ the term you'd like to omit from the result set. Any documents that
 contain this term will not show up in your results. Since this query
 includes two terms, separate the terms with spaces.
 
-.. code-block:: javascript
-
-   const database = client.db("sample_mflix");
-   const collection = database.collection("movies");
-
-   const query = { $text: { $search: "\"star trek\" -\"into darkness\"" } };
-
-   // Return only the `title` of each matched document
-   const projection = {
-      _id: 0,
-      title: 1
-   };
-
-   // find documents based on our query, sort and projection
-   const cursor = collection
-      .find(query)
-      .project(projection);
-
-   // print a message if no documents were found
-   if ((await cursor.count()) == 0) {
-      console.log("No documents found!");
-   }
-   await cursor.forEach(console.dir);
+.. literalinclude:: /code-snippets/crud/startrek.js
+   :language: javascript
+   :start-after: start negation text example
+   :end-before: end negation text example
+   :dedent: 4
 
 Querying with the negated term yields the following documents:
 
@@ -210,33 +161,11 @@ Now that the result set reflects the desired results, you can use the
 text search ``textScore``, accessed using the ``$meta`` operator in the
 query projection, to order the results by relevance:
 
-.. code-block:: javascript
-
-   const database = client.db("sample_mflix");
-   const collection = database.collection("movies");
-
-   const query = { $text: { $search: "\"star trek\"  -\"into darkness\"" } };
-
-   // sort returned documents by descending text relevance score
-   const sort = { score: { $meta: "textScore" } };
-   // Include only the `title` and `score` fields in each returned document
-   const projection = {
-      _id: 0,
-      title: 1,
-      score: { $meta: "textScore" },
-   };
-
-   // find documents based on our query, sort, and projection
-   const cursor = collection
-      .find(query)
-      .sort(sort)
-      .project(projection);
-
-   // print a message if no documents were found
-   if ((await cursor.count()) == 0) {
-      console.log("No documents found!");
-   }
-   await cursor.forEach(console.dir);
+.. literalinclude:: /code-snippets/crud/startrek.js
+   :language: javascript
+   :start-after: start relevance text example
+   :end-before: end relevance text example
+   :dedent: 4
 
 Querying in this way returns the following documents in the following
 order. In general, text relevance increases as a string matches more
